manual client added

+++ version3/client.c
@@ -8,15 +8,22 @@
 #include <unistd.h>
 
 
-
+int flag=20;
 int proc;
+
 void signal_handler(int signum){
-    if(signum==2){
-        printf("SIGINT\n");
-        kill(proc,2);
-    }else if(signum==3){
-        printf("SIGQUIT\n");
-        kill(proc,3);
+    flag--;
+    if(signum==SIGTSTP){
+        puts("EXIT");
+        exit(1);
+    }else{
+        if(signum==2){
+            printf("SIGINT\n");
+            kill(proc,2);
+        }else if(signum==3){
+            printf("SIGQUIT\n");
+            kill(proc,3);
+        }
     }
 }
 
@@ -24,10 +31,18 @@ int main(){
     
     int p=getpid();
     printf("Process:: %d\n\n",p); 
-    int flag=20;
+    
     int seg=access_segment();
     read_values(seg,&proc);
     srand(time(NULL));
+
+    if (signal(SIGQUIT, signal_handler) == SIG_ERR){
+        perror("can't catch SIGQUIT\n");}
+    if (signal(SIGINT, signal_handler) == SIG_ERR){
+        perror("can't catch SIGINT\n");}
+    if (signal(SIGTSTP, signal_handler) == SIG_ERR){
+        perror("can't catch SIGINT\n");}
+
 
     int signum=0;
 
@@ -36,7 +51,6 @@ int main(){
         printf("Signal %d:: ",flag);
         signum=rand()%2+2;
         signal_handler(signum);
-        flag--;
     }
     
     printf("END of CLIENT\n");
